+++ src/main/php/filter/HttpUriFilter.php
@@ -52,7 +52,7 @@ class HttpUriFilter implements Filter
         try {
             $httpUri = HttpUri::fromString($param->getValue());
         } catch (MalformedUriException $murle) {
-            $param->addErrorWithId('HTTP_URI_INCORRECT');
+            $param->addError('HTTP_URI_INCORRECT');
             return null;
         }
 
@@ -61,7 +61,7 @@ class HttpUriFilter implements Filter
         }
 
         if (true === $this->enforceDnsRecord && !$httpUri->hasDnsRecord()) {
-            $param->addErrorWithId('HTTP_URI_NOT_AVAILABLE');
+            $param->addError('HTTP_URI_NOT_AVAILABLE');
             return null;
         }